refactor: separated custom exceptions to own file
